backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1
google-generativeai
structlog==24.1.0
python-jose[cryptography]==3.3.0
alembic==1.13.1
